Changed yt-dlp command to execute with slightly different options
The yt-dlp command was simplified, and now what it attempts to do is download the best available version of merged audio and video available from youtube. By default, yt-dlp will try to download the best version of the video and best version of the audio separately, then merge them. However, at least on my system, yt-dlp is having trouble downloading videos that way, and the downloads themselves are extremely slow. For now, this version of the script works fine and will download your clips at a decent rate, but they'll probably only be decent quality.